Added javadoc to controller and service layers

+++ src/main/java/com/linkshortener/service/LinkService.java
@@ -14,6 +14,18 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.util.*;
 
+/**
+ * This class consist of methods that make business logic of creation,
+ * retrieving, removing of links. Have possibility to add link to global
+ * user with username anonymousUser to then get it by not authenticated
+ * user. Authenticated user can do the same and also remove the link.
+ *
+ * @author Bohdan Koval
+ * @see LinkDao
+ * @see UserDao
+ * @see Link
+ * @see User
+ */
 @Service
 @Transactional(readOnly = true)
 public class LinkService {
@@ -26,6 +38,12 @@ public class LinkService {
         this.userDao = userDao;
     }
 
+    /**
+     * Returns link by id if found, if not found in current user than returns
+     * empty optional. If username not found throw UsernameNotFoundException.
+     * @param id the id of link to return
+     * @return optional of link, if not found - empty
+     */
     public Optional<Link> getLinkById(long id) {
         Optional<Link> optionalLink = linkDao.get(id);
         Optional<User> optionalUser = getUserFromAuthContext();
@@ -37,6 +55,12 @@ public class LinkService {
         return Optional.empty();
     }
 
+    /**
+     * Returns link by alias if found, if not found in current user than returns
+     * empty optional. If username not found throw UsernameNotFoundException.
+     * @param alias the id of link to return
+     * @return optional of link, if not found - empty
+     */
     public Optional<Link> getLinkByAlias(String alias) {
         Optional<Link> optionalLink = linkDao.getLinkByAlias(alias);
         Optional<User> optionalUser = getUserFromAuthContext();
@@ -48,6 +72,11 @@ public class LinkService {
         return Optional.empty();
     }
 
+    /**
+     * Returns all links of current user. If user is not authenticated, nothing
+     * happens. If username not found throw UsernameNotFoundException.
+     * @return all link of current user
+     */
     public List<Link> getAllLinks() {
         Optional<User> optionalUser = getUserFromAuthContext();
         List<Link> links = new ArrayList<>();
@@ -64,6 +93,11 @@ public class LinkService {
         return links;
     }
 
+    /**
+     * Adds link, sets user to current user. If user not authenticated, link
+     * adds to anonymousUser. If username not found throw UsernameNotFoundException.
+     * @param link the link to add to current user
+     */
     @Transactional
     public void addLink(Link link) {
         Optional<User> optionalUser = getUserFromAuthContext();
@@ -75,6 +109,11 @@ public class LinkService {
         }
     }
 
+    /**
+     * Removes link by id, if link found in current user. If user not
+     * authenticated, nothing happens. If username not found throw UsernameNotFoundException.
+     * @param id the id of link to remove
+     */
     @Transactional
     public void removeLink(long id) {
         Optional<User> optionalUser = getUserFromAuthContext();
@@ -85,11 +124,15 @@ public class LinkService {
         }
     }
 
+    /**
+     * Removes all his links, if user is authenticated . If user not authenticated, nothing
+     * happens. If username not found throw UsernameNotFoundException.
+     */
     @Transactional
     public void removeAllLinks() {
         Optional<User> optionalUser = getUserFromAuthContext();
 
-        if (optionalUser.isPresent()) {
+        if (optionalUser.isPresent() && !optionalUser.get().getEmail().equals("anonymousUser")) {
             linkDao.deleteAllByUserId(optionalUser.get().getId());
         }
     }
@@ -110,12 +153,8 @@ public class LinkService {
     }
 
     private boolean isUsersLink(Optional<User> user, Optional<Link> link) {
-        if (link.isPresent()
+        return link.isPresent()
                 && user.isPresent()
-                && Objects.equals(link.get().getUser().getId(), user.get().getId())) {
-            return true;
-        }
-
-        return false;
+                && Objects.equals(link.get().getUser().getId(), user.get().getId());
     }
 }
